Update jeep-can-dash.ino
Removed melexis functions, needs a full rewrite. This will be in a separate program. Added buffering to analog temperature reads. Added info about CAN info available.

https://docs.google.com/spreadsheets/d/1-J7CbgGefj09fzZcbyEStubXSnJ0lc4kMVt5LscqHV0/edit?usp=sharing

changed all 'put' methods to 'set' because for some reason I decided to not follow the standard of using setters and getters to retrieve values.

Cleaned up the CAN connection routine, added an interrupt to prevent the CAN chip from running out of buffer.

added code to emulate the CAN addresses broadcast by the stock cluster

Also, i decided I am not implementing analog reading of the backlight intensity for now and the status of the security lamp. Removed the code for that.